+++ R/buildEdgeMatrix.R
@@ -52,13 +52,13 @@ buildWijMatrix <- function(x,
 buildWijMatrix.TsparseMatrix <- function(x,
 																				 threads = NULL,
 																	 perplexity = 50) {
-	wij <- referenceWij(x@j, x@i, x@x^2, threads, perplexity);
+	wij <- referenceWij(x@j, x@i, x@x^2, as.integer(threads), perplexity);
 	return(wij)
 }
 #' @export
 #' @rdname buildWijMatrix
 buildWijMatrix.CsparseMatrix <- function(x, threads = NULL, perplexity = 50) {
 	is <- rep(0:(ncol(x) - 1), diff(x@p))
-  wij <- referenceWij(is, x@i, x@x^2, threads, perplexity)
+  wij <- referenceWij(is, x@i, x@x^2, as.integer(threads), perplexity)
   return(wij)
 }